tls: refactor tls mux listener package (#2161)
