+++ transport/internal/tls/muxlistener/conn_sniffer_test.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package tlsmux
+package muxlistener
 
 import (
 	"bytes"
